Separate code into other files
